renamed incomprehensible names of methods and classes, added seperate…

+++ src/main/java/com/exadel/sandbox/role/repository/RoleRepository.java
@@ -3,6 +3,5 @@ package com.exadel.sandbox.role.repository;
 import com.exadel.sandbox.role.entity.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoleRepository extends JpaRepository<Role, Integer> {
-
+public interface RoleRepository extends JpaRepository<Role, Long> {
 }
